+++ index.d.ts
@@ -1,14 +1,20 @@
 import type { Readable, Writable } from "svelte/store";
 
-// Source of `Stores` and `StoresValues`:
-// https://github.com/sveltejs/svelte/blob/master/src/runtime/store/index.ts
+/** The minimal requirements of the
+ * [writable store contract](https://svelte.dev/docs#component-format-script-4-prefix-stores-with-$-to-access-their-values-store-contract).
+ */
+type MinimalWritable<T> = Pick<Writable<T>, "set" | "subscribe">;
 
-/** One or more `Readable`s. */
-type Stores = Readable<any> | [Readable<any>, ...Array<Readable<any>>] | Array<Readable<any>>;
+/** Stores that may be used as origins. */
+type Stores = MinimalWritable<any> | [Readable<any>, ...Array<Readable<any>>] | Array<Readable<any>>;
 
-/** One or more values from `Readable` stores. */
+/** Values retrieved from origin stores. */
 type StoresValues<T> = T extends Readable<infer U> ? U :
     { [K in keyof T]: T[K] extends Readable<infer U> ? U : never };
+
+/** Values sent to origin stores. */
+type SetValues<T> = T extends MinimalWritable<infer U> ? U :
+    { [K in keyof T]: T[K] extends MinimalWritable <infer U> ? U : never };
 
 /**
  * Create a store similar to [Svelte's `derived`](https://svelte.dev/docs#derived), but which
@@ -30,42 +36,42 @@ type StoresValues<T> = T extends Readable<infer U> ? U :
 export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>) => T,
-    reflect: (reflecting: T) => StoresValues<S>,
+    reflect: (reflecting: T) => SetValues<S>,
     initial?: T
 ): Writable<T>;
 
 export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>, set: (value: T) => void) => void,
-    reflect: (reflecting: T) => StoresValues<S>,
+    reflect: (reflecting: T) => SetValues<S>,
     initial?: T
 ): Writable<T>;
 
 export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>) => T,
-    reflect: (reflecting: T, set: (value: StoresValues<S>) => void) => void,
+    reflect: (reflecting: T, set: (value: SetValues<S>) => void) => void,
     initial?: T
 ): Writable<T>;
 
 export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>, set: (value: T) => void) => void,
-    reflect: (reflecting: T, set: (value: StoresValues<S>) => void) => void,
+    reflect: (reflecting: T, set: (value: SetValues<S>) => void) => void,
     initial?: T
 ): Writable<T>;
 
 export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>) => T,
-    reflect: { withOld: (reflecting: T, old: StoresValues<S>) => StoresValues<S> },
+    reflect: { withOld: (reflecting: T, old: StoresValues<S>) => SetValues<S> },
     initial?: T
 ): Writable<T>;
 
 export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>, set: (value: T) => void) => void,
-    reflect: { withOld: (reflecting: T, old: StoresValues<S>) => StoresValues<S> },
+    reflect: { withOld: (reflecting: T, old: StoresValues<S>) => SetValues<S> },
     initial?: T
 ): Writable<T>;
 
@@ -73,7 +79,7 @@ export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>) => T,
     reflect: {
-        withOld: (reflecting: T, old: StoresValues<S>, set: (value: StoresValues<S>) => void) => void
+        withOld: (reflecting: T, old: StoresValues<S>, set: (value: SetValues<S>) => void) => void
     },
     initial?: T
 ): Writable<T>;
@@ -82,7 +88,7 @@ export default function writableDerived<S extends Stores, T>(
     origins: S,
     derive: (values: StoresValues<S>, set: (value: T) => void) => void,
     reflect: {
-        withOld: (reflecting: T, old: StoresValues<S>, set: (value: StoresValues<S>) => void) => void
+        withOld: (reflecting: T, old: StoresValues<S>, set: (value: SetValues<S>) => void) => void
     },
     initial?: T
 ): Writable<T>;
@@ -100,12 +106,12 @@ export { writableDerived };
  * @returns A writable store.
  */
 export function propertyStore<O extends object, K extends keyof O>(
-    origin: Writable<O>,
+    origin: MinimalWritable<O>,
     propName: K | [K]
 ): Writable<O[K]>;
 
 export function propertyStore<O extends object, K1 extends keyof O, K2 extends keyof O[K1]>(
-    origin: Writable<O>,
+    origin: MinimalWritable<O>,
     propName: [K1, K2]
 ): Writable<O[K1][K2]>;
 
@@ -115,7 +121,7 @@ export function propertyStore<
     K2 extends keyof O[K1],
     K3 extends keyof O[K1][K2]
 >(
-    origin: Writable<O>,
+    origin: MinimalWritable<O>,
     propName: [K1, K2, K3]
 ): Writable<O[K1][K2][K3]>;
 
@@ -126,11 +132,11 @@ export function propertyStore<
     K3 extends keyof O[K1][K2],
     K4 extends keyof O[K1][K2][K3]
 >(
-    origin: Writable<O>,
+    origin: MinimalWritable<O>,
     propName: [K1, K2, K3, K4]
 ): Writable<O[K1][K2][K3][K4]>;
 
 export function propertyStore(
-    origin: Writable<object>,
+    origin: MinimalWritable<object>,
     propName: string | number | symbol | Array<string | number | symbol>
 ): Writable<any>;